Move everything around again because Python's import mechanism is horrible
