+++ datascrubber/tasks/__init__.py
@@ -1,0 +1,3 @@
+from .whitehall import scrub_whitehall
+from .email_alert_api import scrub_email_alert_api
+from .publishing_api import scrub_publishing_api
