+++ datascrubber/tasks/whitehall.py
@@ -1,9 +1,7 @@
 import logging
-import os.path
 
 
 # SQL copied over from github.com/alphagov/whitehall/script/scrub-database
-
 def scrub_whitehall(cursor):
     logger = logging.getLogger('scrub_whitehall')
 
@@ -155,26 +153,3 @@ def scrub_whitehall(cursor):
     cursor.execute(sql, params=(lorem_ipsum_paragraphs,))
     logger.info('Rows affected: %d', cursor.rowcount)
 
-
-def scrub_email_alert_api(cursor):
-    logger = logging.getLogger('scrub_email_alert_api')
-    sql_script = os.path.join(
-        os.path.dirname(__file__),
-        'sql', 'scrub_email_alert_api.sql'
-    )
-    logger.info("Loading SQL from %s", sql_script)
-    sql = open(sql_script, 'r').read()
-    logger.info("Executing SQL (%d lines) ...", len(sql.split('\n')))
-    cursor.execute(sql)
-
-
-def scrub_publishing_api(cursor):
-    logger = logging.getLogger('scrub_publishing_api')
-    sql_script = os.path.join(
-        os.path.dirname(__file__),
-        'sql', 'scrub_publishing_api.sql'
-    )
-    logger.info("Loading SQL from %s", sql_script)
-    sql = open(sql_script, 'r').read()
-    logger.info("Executing SQL (%d lines) ...", len(sql.split('\n')))
-    cursor.execute(sql)
